src/related_work_implem/requirements.txt
IPy==0.83
requests==2.13.0
py-radix==0.9.7
bs4==0.0.1